pyproject.toml
[project]
name = "mdmr"
version = "0.1.0"
description = "MDMR: Maximize Diversity Minimize Redundancy"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "numpy>=1.24",
]

[project.optional-dependencies]
paper = [
  "cellpose==3.1.1",
  "numpy==1.24",
  "PyYAML==6.0.2",
  "scikit_learn==1.2.2",
  "scikit-image==0.21.0",
  "tifffile==2025.9.9",
  "typer==0.17.4",
  "dppy==0.3.3",              # for the DPP selector method
  "FADS==1.0.1",              # for the FADS selector method
  "apricot-select==0.6.1",    # for the Facility Location selector method
  "pandas==2.0.3",            # for the typiclust method
  # "faiss-gpu==1.7.2"        # for the typiclust method. does not install with pip, use conda
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
  "mdmr",
  "mdmr_paper",
]

[tool.setuptools.package-dir]
mdmr = "mdmr"
mdmr_paper = "mdmr-paper/src/mdmr_paper"

[project.scripts]
# Optional: CLI entry points (available after installing with [paper])
mrmd-fine-tune = "mdmr_paper.scripts.fine_tuning:app"
mrmd-eval   = "mdmr_paper.scripts.eval_pretrained:app"
mrmd-fs     = "mdmr_paper.scripts.train_fully_supervised:app"
